requirements.txt
anyio==3.6.2
async-generator==1.10
attrs==22.2.0
bcrypt==4.1.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
ecdsa==0.18.0
exceptiongroup==1.1.1
fastapi==0.95.0
greenlet==3.0.3
h11==0.14.0
idna==3.4
outcome==1.2.0
passlib==1.7.4
psycopg2-binary==2.9.5
pyasn1==0.5.1
pycparser==2.21
pydantic==1.10.7
PySocks==1.7.1
python-jose==3.3.0
python-multipart==0.0.6
requests==2.28.2
rsa==4.9
selenium==4.8.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.25
starlette==0.26.1
trio==0.22.0
trio-websocket==0.10.2
urllib3==1.26.15
uvicorn==0.21.1
wsproto==1.2.0
google-generativeai